MonoBehaviour Coroutine 초안 개선판
추가적으로 CanResume, UpdateCoroutine 등 추가. 이후 엔진 생애 주기 중 Scene의 GameLogic 부분 상에 얹을 예정

+++ DUOLGameEngine/include/DUOLGameEngine/Util/Coroutine/Coroutine.h
@@ -33,7 +33,7 @@ namespace DUOLGameEngine
 		DEFINE_DEFAULT_COPY_MOVE(Coroutine)
 
 	public:
-		virtual void UpdateInstruction(double deltaTime) override;
+		virtual void UpdateInstruction(float deltaTime) override;
 
 		virtual bool CanResume() override;
